+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this module are documented in this file.
 
 valr-api-node module adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2021-10-28
+* Added: methods `getApiKeyInfo()`, `stopLimitOrder()` and `batchOrders()` 
+* Fixed: dependencies updated
+
 ## [1.3.1] - 2020-09-28
 * Fixed: dependencies updated
 
 ## [1.3.0] - 2020-07-12
 * Added: optional parameter `full` for method `getOrderBook`
-* Added: method `getStatus`
+* Added: method `getStatus()`
 
 ## [1.2.0] - 2020-01-11
 
@@ -19,7 +23,7 @@ valr-api-node module adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.2] - 2020-01-11
 
 * Fixed: dependencies updated
-* Fixed: documentation for `getOpenOrders` 
+* Fixed: documentation for `getOpenOrders()` 
 
 ## [1.1.1] - 2019-08-13